LayoutTests/fast/mediastream/MediaStream-construct-with-ended-tracks-expected.txt


PASS typeof mediaStream.active is 'boolean'
PASS mediaStream.active is true
videoTrack = mediaStream.getVideoTracks()[0]
audioTrack = mediaStream.getAudioTracks()[0]
PASS videoTrack.readyState is not "ended"
PASS audioTrack.readyState is not "ended"
PASS videoTrack.readyState is "ended"
PASS audioTrack.readyState is "ended"
PASS typeof newMediaStream.active is 'boolean'
PASS newMediaStream.active is false
PASS mediaStream.id is not newMediaStream.id
PASS typeof mediaStreamCopy.active is 'boolean'
PASS mediaStreamCopy.active is false
PASS mediaStreamCopy.id is not newMediaStream.id
PASS successfullyParsed is true

TEST COMPLETE
